+++ cmd/internal/cryptox/package.go
@@ -1,0 +1,1 @@
+package cryptox // import "code.cloudfoundry.org/perm/cmd/internal/cryptox"
