+++ cmd/internal/ioutilx/file_or_string.go
