+++ cmd/internal/ioutilx/ioutilxfakes/fake_file_info.go
