+++ cmd/internal/ioutilx/package.go
@@ -1,0 +1,1 @@
+package ioutilx // import "code.cloudfoundry.org/perm/cmd/internal/ioutilx"
