--- pkg/cryptox/package.go
@@ -1,1 +1,0 @@
-package cryptox // import "code.cloudfoundry.org/perm/pkg/cryptox"
