--- pkg/ioutilx/package.go
@@ -1,1 +1,0 @@
-package ioutilx // import "code.cloudfoundry.org/perm/pkg/ioutilx"
